Initial commit of refactoring.  Need to test.
