+++ node_server/http_status_codes.js
@@ -1,0 +1,3 @@
+module.exports.http_status_ok = 200;
+module.exports.http_status_no_contents = 204;
+module.exports.http_status_not_found = 404;
